Changed DTo definisions
