+++ Pop-Grupa-A-Backend/app/main/util/DTO/ComputationAccount/UserDTO.py
@@ -1,0 +1,11 @@
+from flask_restplus import Namespace, fields
+
+
+class UserDto:
+    api = Namespace('user', description='user related operations')
+    user = api.model('user', {
+        'email': fields.String(required=True, description='user email address'),
+        'username': fields.String(required=True, description='user username'),
+        'password': fields.String(required=True, description='user password'),
+        'id': fields.Integer(description='user Identifier')
+    })
